Rename data_handling folder to telemetry
