+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ ci:
 
 repos:
 -   repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: 'v0.8.6'
+    rev: 'v0.9.2'
     hooks:
     - id: ruff  # Runs the Ruff linter
       name: ruff linter
