+++ airbrakes/telemetry/logger.py
@@ -8,8 +8,8 @@ from collections import deque
 from pathlib import Path
 from typing import Any, Literal
 
-from airbrakes.data_handling.packets.apogee_predictor_data_packet import ApogeePredictorDataPacket
-from airbrakes.data_handling.packets.servo_data_packet import ServoDataPacket
+from airbrakes.telemetry.packets.apogee_predictor_data_packet import ApogeePredictorDataPacket
+from airbrakes.telemetry.packets.servo_data_packet import ServoDataPacket
 from airbrakes.utils import modify_multiprocessing_queue_windows
 
 if sys.platform != "win32":
@@ -26,10 +26,10 @@ from airbrakes.constants import (
     MAX_GET_TIMEOUT_SECONDS,
     STOP_SIGNAL,
 )
-from airbrakes.data_handling.packets.context_data_packet import ContextDataPacket
-from airbrakes.data_handling.packets.imu_data_packet import EstimatedDataPacket, IMUDataPacket
-from airbrakes.data_handling.packets.logger_data_packet import LoggerDataPacket
-from airbrakes.data_handling.packets.processor_data_packet import ProcessorDataPacket
+from airbrakes.telemetry.packets.context_data_packet import ContextDataPacket
+from airbrakes.telemetry.packets.imu_data_packet import EstimatedDataPacket, IMUDataPacket
+from airbrakes.telemetry.packets.logger_data_packet import LoggerDataPacket
+from airbrakes.telemetry.packets.processor_data_packet import ProcessorDataPacket
 
 
 class Logger:
