+++ airbrakes/telemetry/packets/imu_data_packet.py
