+++ scripts/run_logger.py
@@ -3,10 +3,10 @@
 import time
 from collections import deque
 
-from airbrakes.data_handling.data_processor import IMUDataProcessor
+from airbrakes.telemetry.data_processor import IMUDataProcessor
 from airbrakes.constants import TEST_LOGS_PATH
-from airbrakes.data_handling.packets.imu_data_packet import RawDataPacket
-from airbrakes.data_handling.logger import Logger
+from airbrakes.telemetry.packets.imu_data_packet import RawDataPacket
+from airbrakes.telemetry.logger import Logger
 
 
 def main():
