+++ tests/auxil/utils.py
@@ -1,10 +1,10 @@
 """Utility functions used throughout the test suite."""
 
-from airbrakes.data_handling.packets.apogee_predictor_data_packet import ApogeePredictorDataPacket
-from airbrakes.data_handling.packets.context_data_packet import ContextDataPacket
-from airbrakes.data_handling.packets.imu_data_packet import EstimatedDataPacket, RawDataPacket
-from airbrakes.data_handling.packets.processor_data_packet import ProcessorDataPacket
-from airbrakes.data_handling.packets.servo_data_packet import ServoDataPacket
+from airbrakes.telemetry.packets.apogee_predictor_data_packet import ApogeePredictorDataPacket
+from airbrakes.telemetry.packets.context_data_packet import ContextDataPacket
+from airbrakes.telemetry.packets.imu_data_packet import EstimatedDataPacket, RawDataPacket
+from airbrakes.telemetry.packets.processor_data_packet import ProcessorDataPacket
+from airbrakes.telemetry.packets.servo_data_packet import ServoDataPacket
 
 
 def make_est_data_packet(**kwargs) -> EstimatedDataPacket:
