+++ tests/test_components/test_data_processor.py
@@ -8,8 +8,8 @@ import numpy.testing as npt
 import pytest
 import scipy.spatial
 
-from airbrakes.data_handling.data_processor import IMUDataProcessor
-from airbrakes.data_handling.packets.imu_data_packet import EstimatedDataPacket
+from airbrakes.telemetry.data_processor import IMUDataProcessor
+from airbrakes.telemetry.packets.imu_data_packet import EstimatedDataPacket
 from tests.auxil.utils import make_est_data_packet
