+++ tests/test_packets/test_imu_data_packet.py
@@ -1,6 +1,6 @@
 import pytest
 
-from airbrakes.data_handling.packets.imu_data_packet import EstimatedDataPacket, RawDataPacket
+from airbrakes.telemetry.packets.imu_data_packet import EstimatedDataPacket, RawDataPacket
 
 
 class TestEstimatedDataPacket:
